This should just work in all platforms
This trunc() include should work on all platforms, but I am still
a little unsure.  The C++ library called cmath is kind of a wrapper
for the C99 math.h, but it does appear to have trunc() support
outside just math.h trunc().  We can test on VC++ and see.